fixed big int

+++ db/migrate/20131112234426_add_ratings_to_comments.rb
@@ -1,5 +1,5 @@
 class AddRatingsToComments < ActiveRecord::Migration
   def change
-    add_column :comments, :rating, :integer
+    add_column :comments, :rating, :bigint
   end
 end
